+++ EnhanceQoL/NewSettingsTable.lua
@@ -1,6 +1,10 @@
 local addonName, addon = ...
 
 addon.variables.NewVersionTableEQOL = {
+	["EQOL_CombatDungeons"] = true,
 	["EQOL_Quest"] = true,
+	["EQOL_Social"] = true,
+	["EQOL_autoAcceptSummon"] = true,
+	["EQOL_autoReleasePvP"] = true,
 	["EQOL_quickSkipCinematic"] = true,
 }